replace previous interactions code
